Add files via upload
A VERY rough attempt at the website using flask and react. This is the video tutorial I was using:  https://www.youtube.com/watch?v=7LNl2JlZKHA

When I entered the link in my browser to the site, I got an error saying the site could not be reached. Tried some things to fix it but nothing's been working :(